fix complex state + multiple files
